\name{Association.int.plot}
\alias{Association.int.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot Association cumulative distribution function}
\description{
  Plot Association cumulative distribution function
}
\usage{
Association.int.plot(obj, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{list with x and y coordinates for plot}
  \item{\dots}{graphical parameters can be given as arguments to 'plot'}
}
\details{
  This is an internal function called by \code{Association.plot} and is not intended for stand-alone use.
}

\author{Stephen J. Smith \email{smithsj@eastlink.ca}}

\seealso{\code{\link{Association.plot}}}

\keyword{internal}% 
