\name{ECDF.strata.plot}
\alias{ECDF.strata.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot empirical cumulative distribution}
\description{
  Plot empirical cumulative distribution for strata object.
}
\usage{
ECDF.strata.plot(obj, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{list with x and y coordinates for plot}
  \item{\dots}{graphical parameters can be given as arguments to 'plot'}
}
\details{
  This is an internal function called by \code{ECDF.strata} and is not intended for stand-alone use.
}

\author{Stephen J. Smith \email{smithsj@eastlink.ca}}

\seealso{\code{\link{ECDF.strata}}}

\keyword{internal}% 
