\name{pref.ind}
\alias{pref.ind}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal function for association analysis}
\description{
Simple comparison function
  }
\usage{
pref.ind(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{vector}
  \item{y}{vector}
}

\value{
  A vector containing entries of 1 or 2 referring to whether x <= y is false or true.
}

\keyword{internal}% 

